+++ package.json
@@ -112,7 +112,6 @@
   "devDependencies": {
     "@semantic-release/exec": "^5.0.0",
     "@semantic-release/git": "^9.0.0",
-    "@types/graphql": "^14.5.0",
     "@types/jest": "^23.3.2",
     "@types/node": "^10.11.0",
     "@types/w3c-web-usb": "^1.0.4",
