+++ src/eventemitter.ts
@@ -1,11 +1,6 @@
 import { SMap } from "./utils";
-// tslint:disable-next-line: no-submodule-imports
-import { PubSubEngine } from "graphql-subscriptions/dist/pubsub-engine"
-
+import { NEW_LISTENER, REMOVE_LISTENER } from "./constants";
 export type EventHandler = (evt: any) => void;
-
-export const NEW_LISTENER = 'newListener'
-export const REMOVE_LISTENER = 'removeListener'
 
 export class EventEmitter {
     readonly listeners: SMap<EventHandler[]> = {};
@@ -48,41 +43,3 @@ export class EventEmitter {
     }
 }
 
-export interface PubSubOptions {
-    eventEmitter?: EventEmitter;
-}
-
-export class PubSub extends PubSubEngine {
-    protected ee: EventEmitter;
-    private subscriptions: { [key: string]: [string, (...args: any[]) => void] };
-    private subIdCounter: number;
-
-    constructor(options: PubSubOptions = {}) {
-        super();
-        this.ee = options.eventEmitter || new EventEmitter();
-        this.subscriptions = {};
-        this.subIdCounter = 0;
-    }
-
-    public publish(triggerName: string, payload: any): Promise<void> {
-        this.ee.emit(triggerName, payload);
-        return Promise.resolve();
-    }
-
-    public subscribe(triggerName: string, onMessage: EventHandler): Promise<number> {
-        this.ee.addListener(triggerName, onMessage);
-        this.subIdCounter = this.subIdCounter + 1;
-        this.subscriptions[this.subIdCounter] = [triggerName, onMessage];
-
-        return Promise.resolve(this.subIdCounter);
-    }
-
-    public unsubscribe(subId: number) {
-        const subs = this.subscriptions[subId];
-        if (!subs) return;
-
-        const [triggerName, onMessage] = subs;
-        delete this.subscriptions[subId];
-        this.ee.removeListener(triggerName, onMessage);
-    }
-}
