+++ src/jacdac.ts
@@ -4,6 +4,7 @@ export * from './buffer'
 export * from './msgpack'
 export * from './struct'
 export * from './eventemitter'
+export * from './pubsub'
 export * from './packet'
 export * from './device'
 export * from './service'
